refactor: Move debug discovery to a separate class
